Added PostageOption.cs

+++ PackageCalculator/Item.cs
@@ -3,8 +3,8 @@ namespace PackageCalculator;
 public class Item
 {
     public string Name { get; set; }
-    public double Weight { get; set; }
-    public double Length { get; set; }
-    public double Width { get; set; }
-    public double Height { get; set; }
+    public int Weight { get; set; }
+    public int Length { get; set; }
+    public int Width { get; set; }
+    public int Height { get; set; }
 }
